updated the code for derived classes in separete files
