+++ src/main/java/com/padmaja/inheritance/Bicycle.java
@@ -1,0 +1,38 @@
+package com.padmaja.inheritance;
+//import static java.lang.System.out;
+import java.util.Scanner;
+import java.lang.String;
+// base class 
+class Bicycle
+{
+    // the Bicycle class has two fields 
+    public int gear;
+    public int speed;
+
+    // the Bicycle class has one constructor 
+    public Bicycle(int gear, int speed)
+    {
+        this.gear = gear;
+        this.speed = speed;
+    }
+
+    // the Bicycle class has three methods 
+    public void applyBrake(int decrement)
+    {
+        speed -= decrement;
+    }
+
+    public void speedUp(int increment)
+    {
+        speed += increment;
+    }
+
+    // toString() method to print info of Bicycle 
+    public String toString()
+    {
+        return("No of gears are "+gear
+                +"\n"
+                + "speed of bicycle is "+speed);
+    }
+} 
+
