+++ src/main/java/com/padmaja/inheritance/MountainBike.java
@@ -1,0 +1,35 @@
+package com.padmaja.inheritance;
+
+// derived class
+
+class MountainBike extends Bicycle
+{
+
+    // the MountainBike subclass adds one more field
+    public int seatHeight;
+
+    // the MountainBike subclass has one constructor
+    public MountainBike(int gear,int speed,
+                        int startHeight)
+    {
+        // invoking base-class(Bicycle) constructor
+        super(gear, speed);
+        seatHeight = startHeight;
+    }
+
+    // the MountainBike subclass adds one more method
+    public void setHeight(int newValue)
+    {
+        seatHeight = newValue;
+    }
+
+    // overriding toString() method
+    // of Bicycle to print more info
+    @Override
+    public String toString()
+    {
+        return (super.toString()+
+                "\nseat height is "+seatHeight);
+    }
+
+}
