--- src/main/java/com/padmaja/inheritance/PadmajaInheritance.java
@@ -1,10 +1,0 @@
-package com.padmaja.inheritance;
-//import static java.lang.System.out;
-import java.util.Scanner;
-import java.lang.String;
-
-public class PadmajaInheritance {
-}
-
-
-
